Restore basic FIXED subframe generation to overhauled FLAC encoder.
This differs from the Python-based encoder in that it estimates
residual partition sizes rather than generating the partitions
and counting bits.  It's more in line with the reference encoder
so I've update the encoding documentation to match.

+++ src/pcmreader2.c
@@ -124,7 +124,7 @@ pcmr_read2(struct pcm_reader2 *reader,
     struct pcmr_callback2* callback;
     PyObject* string_obj;
     unsigned char* string;
-    unsigned long string_length;
+    Py_ssize_t string_length;
 
     /*make a call to "pcmreader.read(bytes)"
       where "bytes" is set to the proper PCM frame count*/
